+++ .eslintrc.js
@@ -12,4 +12,4 @@ module.exports = {
   },
   rules: {
   }
- }
+}
